Worked on packet sending
I had some spare time so I worked on packet sending and worked on some
packet splitting, this new system I have in 2.0 pre-calculates
everything being sent instead of doing it in real time. This is better
because it allows for the sender to just send the packet as fast as
possible, and also when it reaches the last part of the split packet it
can add a little bit of extra data on there.