Worked on Acknowledge packets more
I worked on Acknowledge packets so they act “smart” I guess you could
say. They are now able to condense and simplify records without needing
anything but the raw records from the user for it to condense/simplify,
neato right?

I also removed the “RangedRecord” class, Record itself can now be
ranged and by default it is not. For a single record do new
Record(index), for a ranged record do new Record(index, endIndex). To
check if it is ranged simply do .isRanged() or see if the endIndex is
smaller than 0.