Progress!
It seems that RakNetSessions are very close to working! I have dome some
tests that have succeeded but they were minimal. I also managed to get
the login sequence working for the server! I also started working the
event system some more, it's kind of awkard but I can try and fix it up
so it seems more natural. There is also a glitch where when a session is
registered into the HashMap a ConcurrentModificationException is thrown.
I will try to fix this tommorow :)