Got ACK's, NACK's, and login working
I think I finally got this working, however I noticed RakNet will bump
the sequence number even if it is simply resending a lost packet. This
raises eyebrows because I don't remember it doing this before. I'm not
sure if this a preference thing or a hardcoded thing put into raknet.
Either way, I need to figure out why it does this if possible.