Clients are almost working!
This version of the client in version 2.0 is much more stable than it's
glitchy and finnicky predescessor. I have also almost got clients to
fully connect to the server (Connecting to the server works, but not
actual login). The client will not send any connected messages because
there is no update method, so I will probably have to make a method in
the client where it will have to be started before any connections can
be made. Either that or it will call the function to start itself
through it's constructor :P