Added some exceptions, started on examples, etc.
I decided to remove the "cancelled" boolean in SessionPreparation and
make a reason to why the connection was canclled mandatory, making sure
ServerOfflineException wasn't thrown as a fallback, which could get
confusing. I caught this when it turned out LBSG has their encryption
enabled, which seems to be something recent. Hoepfully RakNet 9 wasn't
released, I don't want to have to update a crap ton of code again just
as I was about done with JRakNet v2 Protocol 8 :(