Fixed a NullPointerException, Added some cool things
I fixed a NullPointerException that would occur during client
disconnection and I also adding a Warning class and then an onWarning
method, which is called when something that is not necessarily an error
occurs but could cause problems. I also deprecated UseSecurityException
and EncryptionEnabledException because they were not being used at the
moment thanks to the new Warning class that has been implemented. I
still need to implement the RakNetServerExample, Chat example, and then
get the shutdown method for the client working :)