Worked on some stability
I noticed this morning that all the decoders and encoders will actually
print the stack trace if anything goes wrong, even if it's caused by an
outside problem. So I went in and tried to fix it and I also implemented
some new features like always. From what I know everything works but I
still need to go and test.