Fixed #23
I attempted to fix bug #23, I am not sure if it's been fixed entirely
but I know that synchronization was a problem so I added a synchronized
modifier to all calls to methods that modify ConcurrentHashMaps. While I
am not certain that this fixed the the issue. it happened alot less
after I added those synchronization modifiers.