Fixed alot of pesky bugs and added a neato feature
In this new verison 2.5, I actually fixed #23, and I also fixed #26 and
#27! I also added a new feature requested in #25 so you can add and
remove new MaximumTransferUnit values to work with using the
RakNetClient. By default, the client adds the machines MTU with 2
retries and then what MCPE uses. However, this can be modified and they
can be removed easily to use something when connecting to a server :)