Fix #38 and update to 2.6.5
Bug #38 has been fixed, it was a small bug, but it could be pretty
pesky! A small sacrifice had to be made though, from now on all client
and server bundles have to manually set the listener to themselves.
However it only takes one line of code to do it, all that is needed is
for the bundle to call .setListenerSelf() before it is started. You can
even do .setListenerSelf().start() and it would work just fine!