Fixed #45, #47, and #48
I fixed those little annoying glitches and I have almost implemeneted
#42, however I should have it done tommorow. And while I was at it, I
added the ability for JRakNet servers and clients to see if they were
connectede to a server/client that is run by JRakNet! This can be
checked in RakNetSession, DiscoveredServer, and ServerPing. I managed to
implement this where it also did not affect the actual protocol either,
so regular RakLib, JRakLib, JRakLibPlus, and RakNet servers/clients work
fine with it!