Fixed broken split packets
I'm not sure why they would not work using the "this". It most likely
has something to do with the reader position. But what's even weirider
is taht encoding works anyways without doing what has to be done in
encode. Encoding still works if you do what is done in decode with "new
Packet(this.buffer());" so I put it there anyways for consistency
reasons. I also added an important not to the EncapsulatedPacket
ackRecord field to warn users when using it.