Added ability to have multiple connection types
I will not push this out as an update just yet. First, I am going to
attempt to see if the creators of RakNet, JRakLib, and JRakLib+ are
willing to implement this functionality. Even if they dont I will keep
this code in the way it is as it will not make a difference either way
:package: make a difference either way :P