Removed JRakLib from ConnectionType
As I was addding support for connection types in JRakLib, I noticed that
it was on an outdated version of the protocol (Version 7), meaing that
adding it would be completely useless. So, I cancelled support for it
and tommorow I will add it for JRakLibPlus as I have permission from
Jython234 to do so.