Added missing packet field
I noticed that the MCPE client was sending a weird value I had never
seen before and was causing problems with CRakNet++. I have bumped the
version and updated the code so it works properly with this new code.
However, I will not push the update until tommorow after another commit
has been made addressing the two questions in the RakNetServer class at
lines 540 and 542-543.