Added comment to RakNetServer
This is just the addition of a comment for me later in case I need to come back to it. I also removed a useless/dead line of code, namely the "continue this loop if the session size is zero" as it did nothing. I will not rebuild 2.9.3 as this small change makes no behavior in the API change.