Fixed #78 and updated to 2.9.4
This is a relatively small change, so much so you might not even need to update. However, a user has found a small bug where an IndexOutOfBoundsExceptionw would occur, rather than the RakNetPacket catching it itself and throwing an IllegalArgumentException stating the buffer given to it was invalid. However, since this was caught and handled by the RakNetServerHandler nothing stopped the server from running, but rather a stack trace was printed making it seem like a problem server side (rather than a client sending bad data) had occurred. As a result, I have decided to make it so the onHandlerException() default methods do not print a stack trace anymore as it can cause confusion especially if a client is programmed incorrectly as it will make it look like the server is broken rather than the client.