Fixed POM
So apparently Eclipse was being stupid and told me there was an error in the POM when there wasn't. As a result, me and two other guys spent the entire day trying to figure out the error. Turned out, I had misplaced a tag. My entire day was defined by a misplaced tag :/