Fixed NullPointerException in MessageIdentifier
There seems to be a bug where the currentName variable will be null on occassion even though packetNames contains a key for it. I'm not sure what exactly causes it. It seems to be some sort of timing error or conflict between clients if they start too close together. This simple (currentName != null) check fixes everything though.