Added support for loopback exemptions
I have added a way for users to easily add loopback exemptions to their softwares automatically. This way, their players will be able to connect to their game servers even if they are running on the same machine as they are on Windows 10 edition. This code might be seen as hacky and looked down upon. But hey, I'm not the one over at Microsoft that thought it would be a good idea to block local loopback connections.

Also bumped version to 2.10.0