Made some small changes
I decided to make a few changes before releasing version 2.10.3. The main changes are that I moved all the code from RakNetUtils and ArrayUtils to the RakNet class and changed the com.whirvis.jraknet.utils.map class to com.whirvis.jraknet.map.

I also made a small improvement when it comes to negotiating the MTU to use during connection: During connection, it is possible the server will respond with an MTU higher than the client was originally going to use (which is what caused bug #95) earlier. At first, I fixed this by simply ignoring this during connection unless an smaller MTU is requested by the server. From now on, it goes like this: If the server requests a higher MTU than the client was originally going to go with, the client will use that if and only if that MTU is not higher than the highest maximum transfer unit provided earlier before connection was attempted. If the server requests a lower MTU than what the client was originally going to go with, then the client will go with that. Also, when responding to OpenConnectionRequestOne, the server will now use its maximum transfer unit when responding with OpenConnectionResponseOne.