FIxed unreliable packets, finished JavaDocs
JRakNet v2.11.0 is nearly finished! Unreliable packets work again, and the chat client and server seem to function perfectly once more. All I have yet to do is determine whether or not listener methods for server, client, and discovery system should run on the same thread or on another thread.