Release v2.12.0
This is the last version I plan to relesae for JRakNet version 2. The only versions I will continue releasing for version 2 will simply be bugfixes.

This is not the end of JRakNet though. The next time I work on JRakNet I will begin work immediately on the third version, which aims to improve the API through chaning major parts of its current framework.

Like I said earlier, version 2 will still be supported. If any bugs are found and reported, they will still be fixed.